Add Ŋŋ
This adds the codepoints for Eng and eng (as well as the glyph Eng.sc), as requested in #64.
While Ŋŋ are not in any Adobe charset below AL-5, the introduction of the new Kazakh Latin Alphabet makes this request a bit more relevant.

Note: this is adding the N-like version of the capital Eng.
The arch-like version has to be added in a future update focused on African languages.