Add dummy contents.plist file to all layers
The glyphs.background and glyphs.com.adobe.type.processedglyphs layers of master UFOs are excluded from this repository, because they create lots of noise.
Unfortunately, it is a catch-22 to keep the repo clean:
- if the glyphs.xxx folders are .gitignored completely, the layercontents.plist file will cause complaints
- if glyphs.xxx folders don’t at least contain a stub contents. plist file, command line tools won’t be able to read the UFOs

This commit is force-adding the contents.plist for each master layer (although .gitignored), hoping that the problem will be solved in this manner.